requirements.txt
Django==5.1
djangorestframework==3.15.2
django-debug-toolbar==4.4.6
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
python-dotenv==1.0.1
telebot==0.0.5
